Добавляет правки по модулям
